Move to System.Web.Adapters namespace
