+++ src/SystemWebAdapters/samples/MvcCoreApp/Controllers/DataController.cs
@@ -1,4 +1,5 @@
 using System.Web;
+using System.Web.Adapters;
 using ClassLibrary;
 using Microsoft.AspNetCore.Mvc;
