+++ src/SystemWebAdapters/samples/MvcCoreApp/Program.cs
@@ -1,4 +1,4 @@
-using System.Web;
+using System.Web.Adapters;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -22,13 +22,6 @@ app.UseStaticFiles();
 app.UseRouting();
 
 app.UseAuthorization();
-
-app.UseSystemWebAdapters();
-
-app.UseEndpoints(endpoints =>
-{
-    endpoints.MapControllers();
-});
 
 app.MapControllerRoute(
     name: "default",
