+++ src/SystemWebAdapters/src/Adapters/BufferRequestStreamAttribute.cs
@@ -1,9 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Web.Metadata;
-
-namespace System.Web;
+namespace System.Web.Adapters;
 
 [AttributeUsage(AttributeTargets.Method | AttributeTargets.Class)]
 public class BufferRequestStreamAttribute : Attribute, IBufferRequestStreamMetadata
