+++ src/SystemWebAdapters/src/Adapters/IBufferRequestStreamMetadata.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace System.Web.Metadata;
+namespace System.Web.Adapters;
 
 public interface IBufferRequestStreamMetadata
 {
