+++ src/SystemWebAdapters/src/Internal/StringValuesNameValueCollection.cs
@@ -2,8 +2,6 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Linq;
-using System.Runtime.Serialization;
-using Microsoft.AspNetCore.Http.Features;
 using Microsoft.Extensions.Primitives;
 
 namespace System.Web.Internal
