+++ src/SystemWebAdapters/test/Adapters/HttpContextAdapterExtensionsTests.cs
@@ -4,7 +4,7 @@
 using Microsoft.AspNetCore.Http;
 using Xunit;
 
-namespace System.Web
+namespace System.Web.Adapters
 {
     public class HttpContextAdapterExtensionsTests
     {
